fixed template

+++ modrep_default/template/test/test_generate.c
@@ -63,8 +63,8 @@ int generate_input_signal(void *in, int *lengths)
 
 	for (i=0;i<block_length;i++) {
 #ifdef GENERATE_COMPLEX
-		__real__ input[i] = (i+offset)%(block_length);
-		__imag__ input[i] = (block_length-i-1+offset)%(block_length);
+		__real__ input[i] = (float) ((i+offset)%(block_length));
+		__imag__ input[i] = (float) ((block_length-i-1+offset)%(block_length));
 #else
 		input[i] = (i+offset)%(block_length);
 #endif
